move handleIgnore to lib/method/helpers
