+++ app/views/RoomMembersView/helpers.ts
@@ -217,20 +217,6 @@ export const handleRemoveUserFromRoom = async (
 	}
 };
 
-export const handleIgnore = async (selectedUser: TUserModel, ignore: boolean, rid: string) => {
-	try {
-		await Services.ignoreUser({
-			rid,
-			userId: selectedUser._id,
-			ignore
-		});
-		const message = I18n.t(ignore ? 'User_has_been_ignored' : 'User_has_been_unignored');
-		EventEmitter.emit(LISTENER, { message });
-	} catch (e) {
-		log(e);
-	}
-};
-
 export const handleOwner = async (
 	selectedUser: TUserModel,
 	isOwner: boolean,
